fix(DBConnection): overflow in tradeid conversion
As defined in the specification, T_ID is of type NUM(15) and this
requires atol instead of atoi for proper conversion, as in e.g.
`storedproc/pgsql/c/trade_order.c:1338`. This overflow in conversion
leads to runs being constantly invalidated as below:

INVALID RUN : see results/bh/BrokerageHouse_Error.log for transaction details
WARNING:  TradeResultFrame1 num_found <> 1 (actual 0) for trade_id 561636692

With printing the string returned from PQgetvalue that is passed to atoi
here, it gives `2000000087422228`. Plus, atol is used in this line prior
to commit f0a34cc74e2ef94ce62e3a15b702530430d1f776, as in old line 537 of
DBConnection in it.

+++ src/transactions/pgsql/DBConnection.cpp
@@ -570,7 +570,7 @@ CDBConnection::execute(const TMarketFeedFrame1Input *pIn,
 			strncpy(m_TriggeredLimitOrders.symbol, pIn->Entries[i].symbol,
 					cSYMBOL_len);
 			m_TriggeredLimitOrders.trade_id
-					= atoi(PQgetvalue(res, j, i_tr_t_id));
+					= atoll(PQgetvalue(res, j, i_tr_t_id));
 			m_TriggeredLimitOrders.price_quote
 					= atof(PQgetvalue(res, j, i_tr_bid_price));
 			m_TriggeredLimitOrders.trade_qty
